added GIDs

+++ Solution/Network/NetMessage.h
@@ -16,8 +16,8 @@ public:
 
 	StreamType myStream;
 	uint32_t myTimeStamp;
-	uint16_t mySenderID;
-	uint16_t myTargetID;
+	uint32_t mySenderID;
+	uint32_t myTargetID;
 	uint8_t myID;
 protected:
